refactor: 로또 번호 생성 기능 분리
