+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,12 +1,11 @@
-package lotto.domain.lottoticket;
+package lotto.domain;
 
 import java.util.Objects;
 
 public final class LottoNumber {
     public static final int MIN = 1;
     public static final int MAX = 45;
-
-    private static final String NUMBER_RANGE_ERROR = "로또 숫자는 " + MIN + " 이상 " + MAX + " 이하의 숫자만 가능합니다.";
+    private static final String NUMBER_RANGE_ERROR = String.format("로또 숫자는 %s 이상 %s 이하의 숫자만 가능합니다.", MIN, MAX);
 
     private final int value;
 
@@ -16,13 +15,10 @@ public final class LottoNumber {
     }
 
     public static LottoNumber valueOf(int lottoNumber) {
-        return LottoNumbersCache.cache.stream()
-                .filter(cachedLottoNumber -> cachedLottoNumber.isSameNumber(lottoNumber))
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException(NUMBER_RANGE_ERROR));
+        return LottoNumberGenerator.getLottoNumber(lottoNumber);
     }
 
-    private boolean isSameNumber(int lottoNumber) {
+    boolean isSameNumber(int lottoNumber) {
         return value == lottoNumber;
     }
