+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoticket;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
